requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
joblib
flask-mysqldb
flask-mail
python-dotenv
gunicorn
reportlab